% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_matrix.R
\docType{data}
\name{A_matrix}
\alias{A_matrix}
\title{A_matrix}
\source{
R package GAPIT
}
\arguments{
\item{A}{a matrix}
}
\description{
A matrix which is example data from R package GAPIT
}
\keyword{datasets}
