% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDK.R
\name{CDK}
\alias{CDK}
\title{construct different kinship matrix}
\usage{
CDK(geno, maf = NULL, A = NULL, EV = NULL, nPCs = 0, method = "VanRaden1")
}
\arguments{
\item{geno}{a matrix}

\item{maf}{a vector of type numeric}

\item{A}{a matrix}

\item{EV}{a vector of type numeric}

\item{nPCs}{the number of front PCs, factor}

\item{method}{a character, options are VanRaden1, VanRaden2, VanRaden3, Weighted, GEMMA1, GEMMA2, GAB, Genetic Distance, GIBS, KING, HIERFSTAT, one-SNP haplotype, GAPIT1, GAPIT2}
}
\value{
a matrix
}
\description{
\code{G_matrix}use different methods to construct the kinship matrix
}
\examples{
G<-CDK(geno=geno_matrix,method="VanRaden1")
}
