% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_matrix.R
\docType{data}
\name{geno_matrix}
\alias{geno_matrix}
\title{geno_matrix}
\source{
R package GAPIT
}
\arguments{
\item{geno}{a matrix}
}
\description{
geno matrix which is example data from R package GAPIT
}
\keyword{datasets}
